requirements.txt
eyeD3
spotipy
python-dotenv
tqdm
matplotlib